Merged PR 63542: Changed the receiver to enforce isolation by default
Removed the autoClone flag from CreateReceiver methods. Messages are by default cloned upon delivery to the receiver, even when executing synchronously. This may in future be modified by changing the `enforceIsolation` field to false (perhaps via DeliveryPolicy), in which case the runtime will first attempt to pass messages by reference when executing synchronously, and only clone the message if it needs to be queued for asynchronous delivery.