Various runtime changes affecting throttling and synchronous message delivery (#229)
* Merged PR 63474: Fixed throttled delivery policies to be lossless

Also,
- enabled an adjustable throttling queue size threshold.
- modified the HoloLensCapture exporter tool to use throttling with ReplayAll.

* Merged PR 63542: Changed the receiver to enforce isolation by default

Removed the autoClone flag from CreateReceiver methods. Messages are by default cloned upon delivery to the receiver, even when executing synchronously. This may in future be modified by changing the `enforceIsolation` field to false (perhaps via DeliveryPolicy), in which case the runtime will first attempt to pass messages by reference when executing synchronously, and only clone the message if it needs to be queued for asynchronous delivery.